VIDEO: Cast unsigned values to signed before negating in Bink decoder
Because negating an unsigned value doesn't make much sense.

+++ src/video/bink.cpp
@@ -1077,7 +1077,7 @@ void Bink::readMotionValues(VideoFrame &video, Bundle &bundle) {
 		byte v = video.bits->getBits(4);
 
 		if (v) {
-			int sign = -video.bits->getBit();
+			int sign = -((int)video.bits->getBit());
 			v = (v ^ sign) - sign;
 		}
 
@@ -1091,7 +1091,7 @@ void Bink::readMotionValues(VideoFrame &video, Bundle &bundle) {
 		byte v = getHuffmanSymbol(video, bundle.huffman);
 
 		if (v) {
-			int sign = -video.bits->getBit();
+			int sign = -((int)video.bits->getBit());
 			v = (v ^ sign) - sign;
 		}
 
@@ -1209,7 +1209,7 @@ void Bink::readDCS(VideoFrame &video, Bundle &bundle, int startBits, bool hasSig
 
 	int32 v = video.bits->getBits(startBits - (hasSign ? 1 : 0));
 	if (v && hasSign) {
-		int sign = -video.bits->getBit();
+		int sign = -((int)video.bits->getBit());
 		v = (v ^ sign) - sign;
 	}
 
@@ -1226,7 +1226,7 @@ void Bink::readDCS(VideoFrame &video, Bundle &bundle, int startBits, bool hasSig
 			for (uint32 j = 0; j < length2; j++) {
 				int16 v2 = video.bits->getBits(bSize);
 				if (v2) {
-					int sign = -video.bits->getBit();
+					int sign = -((int)video.bits->getBit());
 					v2 = (v2 ^ sign) - sign;
 				}
 
@@ -1315,7 +1315,7 @@ void Bink::readDCTCoeffs(VideoFrame &video, int16 *block, bool isIntra) {
 						} else {
 							t = video.bits->getBits(bits) | mask;
 
-							int sign = -video.bits->getBit();
+							int sign = -((int)video.bits->getBit());
 							t = (t ^ sign) - sign;
 						}
 						block[binkScan[ccoef]] = t;
@@ -1340,7 +1340,7 @@ void Bink::readDCTCoeffs(VideoFrame &video, int16 *block, bool isIntra) {
 				} else {
 					t = video.bits->getBits(bits) | mask;
 
-					int sign = -video.bits->getBit();
+					int sign = -((int)video.bits->getBit());
 					t = (t ^ sign) - sign;
 				}
 				block[binkScan[ccoef]] = t;
@@ -1419,7 +1419,7 @@ void Bink::readResidue(VideoFrame &video, int16 *block, int masksCount) {
 					} else {
 						nzCoeff[nzCoeffCount++] = binkScan[ccoef];
 
-						int sign = -video.bits->getBit();
+						int sign = -((int)video.bits->getBit());
 						block[binkScan[ccoef]] = (mask ^ sign) - sign;
 
 						masksCount--;
@@ -1441,7 +1441,7 @@ void Bink::readResidue(VideoFrame &video, int16 *block, int masksCount) {
 			case 3:
 				nzCoeff[nzCoeffCount++] = binkScan[ccoef];
 
-				int sign = -video.bits->getBit();
+				int sign = -((int)video.bits->getBit());
 				block[binkScan[ccoef]] = (mask ^ sign) - sign;
 
 				coefList[listPos]   = 0;
